# docs/swagger.yaml
definitions:
  dopTypes.CreateRep:
    properties:
      id: {}
    type: object
  dopTypes.ErrRep:
    properties:
      desc:
        type: string
      error_code:
        type: string
      fields:
        additionalProperties:
          type: string
        type: object
    type: object
  dopTypes.PaginatedListRep:
    properties:
      page:
        type: integer
      page_size:
        type: integer
      results: {}
      total_count:
        type: integer
    type: object
...
  entities.PermSt:
    properties:
      app_id:
        type: integer
      code:
        type: string
      dsc:
        type: string
      id:
        type: integer
      is_all:
        type: boolean
      is_system:
        type: boolean
    type: object
...
  entities.RoleCUSt:
    properties:
      code:
        type: string
      name:
        type: string
      perm_ids:
        items:
          type: integer
        type: array
    type: object
  entities.RoleListSt:
    properties:
      code:
        type: string
      id:
        type: integer
      is_system:
        type: boolean
      name:
        type: string
    type: object
  entities.RoleSt:
    properties:
      code:
        type: string
      id:
        type: integer
      is_system:
        type: boolean
      name:
        type: string
      perm_ids:
        items:
          type: integer
        type: array
    type: object
...
  entities.UsrCUSt:
    properties:
      ava:
        type: string
      name:
        type: string
      phone:
        type: string
      role_ids:
        items:
          type: integer
        type: array
    type: object
  entities.UsrListSt:
    properties:
      ava:
        type: string
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      phone:
        type: string
    type: object
...
  entities.UsrSt:
    properties:
      ava:
        type: string
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      perm_ids:
        items:
          type: integer
        type: array
      perms:
        items:
          $ref: '#/definitions/entities.PermSt'
        type: array
      phone:
        type: string
      role_ids:
        items:
          type: integer
        type: array
      roles:
        items:
          $ref: '#/definitions/entities.RoleListSt'
        type: array
    type: object
info:
  contact: {}
paths:
...
  /role:
    get:
      parameters:
      - in: query
        name: code
        type: string
      - in: query
        items:
          type: string
        name: cols
        type: array
      - in: query
        items:
          type: integer
        name: ids
        type: array
      - in: query
        name: is_system
        type: boolean
      - in: query
        name: only_count
        type: boolean
      - in: query
        name: page
        type: integer
      - in: query
        name: page_size
        type: integer
      - in: query
        items:
          type: string
        name: sort
        type: array
      - in: query
        name: sort_name
        type: string
      - in: query
        name: with_total_count
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entities.RoleListSt'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dopTypes.ErrRep'
      tags:
      - role
    post:
      parameters:
      - description: body
        in: body
        name: body
        schema:
          $ref: '#/definitions/entities.RoleCUSt'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dopTypes.CreateRep'
            - properties:
                id:
                  type: integer
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dopTypes.ErrRep'
      tags:
      - role
  /role/:id:
    delete:
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dopTypes.ErrRep'
      tags:
      - role
    get:
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.RoleSt'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dopTypes.ErrRep'
      tags:
      - role
    put:
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      - description: body
        in: body
        name: body
        schema:
          $ref: '#/definitions/entities.RoleCUSt'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dopTypes.ErrRep'
      tags:
      - role
  /usr:
    get:
      parameters:
      - in: query
        items:
          type: string
        name: cols
        type: array
      - in: query
        name: id
        type: integer
      - in: query
        items:
          type: integer
        name: ids
        type: array
      - in: query
        name: only_count
        type: boolean
      - in: query
        name: page
        type: integer
      - in: query
        name: page_size
        type: integer
      - in: query
        name: search
        type: string
      - in: query
        items:
          type: string
        name: sort
        type: array
      - in: query
        name: sort_name
        type: string
      - in: query
        name: with_total_count
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dopTypes.PaginatedListRep'
            - properties:
                results:
                  items:
                    $ref: '#/definitions/entities.UsrListSt'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dopTypes.ErrRep'
      tags:
      - usr
    post:
      parameters:
      - description: body
        in: body
        name: body
        schema:
          $ref: '#/definitions/entities.UsrCUSt'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dopTypes.CreateRep'
            - properties:
                id:
                  type: integer
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dopTypes.ErrRep'
      tags:
      - usr
  /usr/:id:
    delete:
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dopTypes.ErrRep'
      tags:
      - usr
    get:
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.UsrSt'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dopTypes.ErrRep'
      tags:
      - usr
    put:
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      - description: body
        in: body
        name: body
        schema:
          $ref: '#/definitions/entities.UsrCUSt'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dopTypes.ErrRep'
      tags:
      - usr
swagger: "2.0"